lispy-eval: improve condition for deciding overlay
We prioritize using eros if available, but we make
sure not to use it if we're in a clojure mode and cider
is installed.

This fixed an issue where lisp modes that didn't use cider
weren't able to display evaluation results with an overlay
because lispy was passing a string argumet to
cider--display-interactive-eval-result, which now requires
a symbol.